feat:add sending delete account request

+++ Backend/ftn/src/main/java/com/fishingbooker/ftn/service/FishingInstructorServiceImpl.java
@@ -49,7 +49,10 @@ public class FishingInstructorServiceImpl implements FishingInstructorService {
 
     @Override
     public Long delete(Long id) {
-        instructorRepository.deleteById(id);
+
+        if (instructorRepository.existsById(id)){
+            instructorRepository.deleteById(id);
+        }
         return id;
     }
